remove _linux specifics, improve gardenUtils, fix typo ins filename
